[dashboardEnhanced] lazy load actions
