+++ x-pack/platform/plugins/shared/dashboard_enhanced/kibana.jsonc
@@ -22,7 +22,6 @@
       "unifiedSearch"
     ],
     "requiredBundles": [
-      "dashboard",
       "embeddable",
       "embeddableEnhanced",
       "kibanaUtils",
