+++ x-pack/platform/plugins/shared/dashboard_enhanced/public/services/drilldowns/async_module.ts
@@ -1,0 +1,10 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export { FlyoutCreateDrilldownAction } from './actions/flyout_create_drilldown/flyout_create_drilldown';
+export { FlyoutEditDrilldownAction } from './actions/flyout_edit_drilldown/flyout_edit_drilldown';
+export { getLocation } from './embeddable_to_dashboard_drilldown/get_location';
