+++ x-pack/platform/plugins/shared/ml/public/application/components/custom_urls/custom_url_editor/utils.ts
@@ -11,9 +11,8 @@ import { cloneDeep } from 'lodash';
 import type { SerializableRecord } from '@kbn/utility-types';
 import rison from '@kbn/rison';
 import url from 'url';
-import { setStateToKbnUrl } from '@kbn/kibana-utils-plugin/public';
+import { cleanEmptyKeys, setStateToKbnUrl } from '@kbn/kibana-utils-plugin/public';
 import type { DashboardStart } from '@kbn/dashboard-plugin/public';
-import { cleanEmptyKeys } from '@kbn/dashboard-plugin/public';
 import type { DataView, DataViewField } from '@kbn/data-views-plugin/common';
 import type { Filter } from '@kbn/es-query';
 import { isFilterPinned } from '@kbn/es-query';
